+++ app/components/omnibox/directives/search-directive.js
@@ -5,10 +5,18 @@
  * and makes sure the right services are called.
  */
 angular.module('omnibox')
-  .directive('search', ['LocationService', 'ClickFeedbackService', 'MapService', 'State',
-  function (LocationService, ClickFeedbackService, MapService, State) {
+  .directive('search', [
+    'LocationService',
+    'ClickFeedbackService',
+    'MapService',
+    'State',
+    '$timeout',
+  function (LocationService, ClickFeedbackService, MapService, State, $timeout) {
 
   var link = function (scope, element, attrs) {
+
+    // Set focus on search input field.
+    element.children()[0].focus();
 
     /**
      * @description event handler for key presses.
@@ -20,12 +28,14 @@ angular.module('omnibox')
 
       if ($event.target.id === "searchboxinput") {
         // Intercept keyPresses *within* searchbox, do xor prevent stuff from happening
-        if ($event.which === 13) {
-          // User hits [enter] -> do search;
-          scope.search();
+        if ($event.which === 13 && scope.box.content.location) {
+          // User hits [enter];
+          scope.zoomTo(scope.box.content.location[0]);
         } else if ($event.which === 32) {
           // user hits [space] -> prevent anim. start/stop
           $event.originalEvent.stopPropagation();
+        } else if ($event.which === 27) { //esc
+          scope.cleanInput();
         }
       }
     };
@@ -35,20 +45,25 @@ angular.module('omnibox')
      * with the right query and puts in on the scope.
      */
     scope.search = function () {
-      if (scope.geoquery && scope.geoquery.length > 1) {
-        LocationService.search(scope.geoquery)
+      if (scope.geoquery.length > 1) {
+        LocationService.search(scope.geoquery, State.spatial)
           .then(function (response) {
-            scope.geoquery = "";
-            if (response.length !== 0) {
-              scope.box.content.location = {
-                data: response
-              };
+            if (response.status === LocationService.ggStatus.OK) {
+              scope.box.content.location = response.results;
+            }
+            else if (
+              response.status === LocationService.ggStatus.ZERO_RESULTS
+            ) {
+              destroyLocationModel();
+            } else {
+              destroyLocationModel();
+              // Throw error so we can find out about it through sentry.
+              throw new Error(
+                'Geocoder returned with status: ' + resonse.status
+              );
             }
           }
         );
-      }
-      else {
-        scope.geoquery = "";
       }
     };
 
@@ -73,7 +88,6 @@ angular.module('omnibox')
      * (5) - Clear the click feedback.
      */
     scope.cleanInput = function () {
-
       State.box.type = "point";
       scope.geoquery = "";
       scope.box.content = {};
@@ -86,19 +100,10 @@ angular.module('omnibox')
      * @description zooms to search result
      * @param {object} one search result.
      */
-    scope.zoomTo = function (obj) {
-      if (obj.boundingbox) {
-        var southWest = new L.LatLng(obj.boundingbox[0], obj.boundingbox[2]);
-        var northEast = new L.LatLng(obj.boundingbox[1], obj.boundingbox[3]);
-        var bounds = new L.LatLngBounds(southWest, northEast);
-        MapService.fitBounds(bounds);
-      } else {
-        if (window.JS_DEBUG) {
-          throw new Error('Oops, no boundingbox on this result - TODO: show a proper message instead of this console error...');
-        }
-      }
+    scope.zoomTo = function (location) {
       destroyLocationModel();
       scope.cleanInput();
+      LocationService.zoomToResult(location);
     };
 
   };