+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog of lizard-nxt client
 Unreleased (1.4.0) (XXXX-XX-XX)
 -------------------------------
 
+- Geocode while typing.
+
+- Move to first result when hitting enter.
+
 - Use google geocoder instead of mapbox.
 
 - Simulate click on precise geocoding results.